4sem/moins/lab2/requirements.txt
numpy==1.24.2
py-cpuinfo==9.0.0
